llmserver/requirements.txt
fastapi[all]
cdp_langchain
cdp_sdk==0.10.3
firebase_admin==6.6.0
langchain_core==0.3.19
langchain_openai==0.2.8
langgraph==0.2.50
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.32.0
web3